requirements.txt
amqp==5.3.1
asgiref==3.8.1
autopep8==2.3.2
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==25.1
django-redis==5.4.0
django-silk==5.3.2
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.0
djangorestframework-stubs==3.15.3
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
flower==2.0.1
gevent==24.11.1
geventhttpclient==2.3.3
gprof2dot==2024.6.6
greenlet==3.1.1
gunicorn==23.0.0
humanize==4.12.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.4.2
locust==2.33.1
MarkupSafe==3.0.2
msgpack==1.1.0
mysqlclient==2.2.7
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==7.0.0
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.1
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==76.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tornado==6.4.2
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
